update: move like button to own component and with own query
